+++ packages/components/molecules/f-card-with-content/CHANGELOG.md
@@ -4,6 +4,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v0.2.0
+------------------------------
+*November 9, 2021*
+
+### Added
+- Heading
+- Description
+- Primary and secondary buttons
+  - `href`s are passed in via prop like `primary-button-href`.
+  - Emit events like `primary-button-click`.
+- Named slot for icon.
+- Storybook story.
+
+
 v0.1.0
 ------------------------------
 *November 3, 2021*